User authentication

+++ Controllers/ListingController.cs
@@ -1,14 +1,12 @@
 using System;
 using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text.Json;
-using System.Threading.Tasks;
 using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.Extensions.Logging;
 using Karma.Models;
 using Karma.Repositories;
+using Microsoft.AspNetCore.Authorization;
+using System.Security.Claims;
 
 namespace Karma.Controllers
 {
@@ -27,8 +25,15 @@ namespace Karma.Controllers
         }
 
         [HttpPost]
-        public ActionResult Post(Listing listing) {
+        [Authorize]
+        public ActionResult Post(Listing listing)
+        {
+            var random = new Random();
+            listing.Id = random.Next(9999).ToString(); // temp fix for id generation, later this should be assigned in DB.
             listing.DatePublished = DateTime.UtcNow; //temp fix for curr date with form submit
+
+            string userId = User.FindFirst(ClaimTypes.Name)?.Value;
+            listing.OwnerId = userId;
             _listingRepository.AddListing(listing);
             return StatusCode(StatusCodes.Status200OK);
         }
@@ -40,14 +45,38 @@ namespace Karma.Controllers
         }
 
         [HttpGet("{id}")]
-        public Listing GetListingById(string id) 
+        public Listing GetListingById(string id)
         {
             return _listingRepository.GetListingById(id);
         }
 
         [HttpDelete("{id}")]
-        public void DeleteListing(string id) {
+        [Authorize]
+        public IActionResult DeleteListing(string id)
+        {
+            string userId = User.FindFirst(ClaimTypes.Name)?.Value;
+            var listing = _listingRepository.GetListingById(id);
+            if (listing.OwnerId != userId)
+                return Unauthorized();
             _listingRepository.DeleteListingById(id);
+            return Ok();
+        }
+
+        [HttpPut("{id}")]
+        [Authorize]
+        public IActionResult UpdateListing(string id, [FromBody] Listing listing)
+        {
+
+            var old = _listingRepository.GetListingById(listing.Id);
+            if (old == null) return NotFound();
+
+            string userId = User.FindFirst(ClaimTypes.Name)?.Value;
+            if (old.OwnerId != userId) return Unauthorized();
+
+            listing.DatePublished = DateTime.UtcNow; //temp fix for curr date with form submit
+            listing.OwnerId = userId;
+            _listingRepository.UpdateListing(listing);
+            return Ok();
         }
     }
 }
